requirements.txt
asgiref==3.2.10
astroid==2.4.2
attrs==19.3.0
Django==3.0.8
djangorestframework==3.11.0
importlib-metadata==1.7.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
more-itertools==8.4.0
packaging==20.4
pluggy==0.13.1
py==1.9.0
pylint==2.5.3
pylint-django==2.1.0
pylint-plugin-utils==0.6
pyparsing==2.4.7
pytest==5.4.3
pytest-django==3.9.0
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
wcwidth==0.2.5
wrapt==1.12.1
zipp==3.1.0
pandas==1.1.1
numpy==1.19.1
scikit-learn==0.23.2
scipy==1.5.2
AvocadoPricePredictionModel==0.3.7